Change user register endpoint to /user
