+++ src/routes/registerRoutes.js
@@ -1,5 +1,5 @@
 module.exports = async (fastify, opts) => {
   fastify.register(require('./root'), { prefix: '/'})
   fastify.register(require('./park'), { prefix: '/park'})
-  fastify.register(require('./auth'), { prefix: '/auth'})
+  fastify.register(require('./user'), { prefix: '/user'})
 }
